requirements.txt
black>=25.12.0
django>=6.0
django-comments-xtd>=2.10.11
django-constance[redis]>=4.3.4
django-contrib-comments>=2.2.0
django-guardian>=3.2.0
django-htmx>=1.27.0
django-imagekit>=6.0.0
django-meta>=2.5.0
django-simple-captcha>=0.6.3
django-tailwind[cookiecutter,honcho,reload]>=4.4.2
django-watson>=1.6.3
django-widget-tweaks>=1.5.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
faker>=40.1.0
flake8>=7.3.0
isort>=7.0.0
loguru>=0.7.3
markdown>=3.10
bleach>=6.1.0
python-magic>=0.4.27
python-dotenv>=1.0.1
psycopg[binary]>=3.1.8
model-bakery>=1.21.0
pendulum>=3.1.0
pillow>=12.1.0
psutil>=7.2.1
pytest>=9.0.2
pytest-cov>=7.0.0
pytest-django>=4.11.1
pydantic>=2.12.5
dj-database-url>=3.1.0
django-redis>=6.0.0
uvicorn[standard]>=0.40.0
django-environ>=0.12.0
